+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Innmind\Immutable\Sequence::sink()->attempt()`
+
 ## 5.12.0 - 2025-03-19
 
 ### Added